+++ src/app/sites/offers/components/filter-row/filter-combo-group.component.ts
@@ -3,13 +3,5 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'app-filter-combo-group',
   template: '<ng-content></ng-content>',
-  styles: [
-    `
-      :host {
-        display: inline-block;
-        margin-bottom: 15px;
-      }
-    `,
-  ],
 })
 export class FilterComboGroupComponent {}